fix: makes SimpleHttpOperator types customizable for mypy
